feat: Added module to create new user
According with the additional behaviors, I added a module to allow
users to create new users (at least 6 more based on the available
avatars).

Also, I updated the _DATA.js file in order to allow save the new
questions and answers of the new users. If this change was not made,
then it throws an error every time a new question is created by the
new user or if the new user tries to answer a poll.